common types combined in files and dashboard and feed types added
