+++ src/services/reducers/ingredient-details-reducers.ts
@@ -1,19 +1,5 @@
 import { createSlice, PayloadAction } from '@reduxjs/toolkit';
-
-export type TDetails = {
-    _id: string;
-    name: string;
-    type: string;
-    proteins: number;
-    fat: number; 
-    carbohydrates: number;
-    calories: number;
-    price: number;
-    image: string;
-    image_mobile: string;
-    image_large: string;
-    __v: number;
-};
+import { TDetails } from '../../types/ingredient-types';
 
 type TIngredientDetailsInitial = {
     productDetails: {} | TDetails;
