+++ src/types/ingredient-types.ts
@@ -1,0 +1,34 @@
+export type TIngredient = {
+    _id: string;
+    name: string;
+    type: string;
+    proteins: number;
+    fat: number;
+    carbohydrates: number;
+    calories: number;
+    price: number;
+    image: string;
+    image_mobile: string;
+    image_large: string;
+    __v: number;
+    }
+
+export type TDetails = {
+    _id: string;
+    name: string;
+    type: string;
+    proteins: number;
+    fat: number; 
+    carbohydrates: number;
+    calories: number;
+    price: number;
+    image: string;
+    image_mobile: string;
+    image_large: string;
+    __v: number;
+};
+
+export type TIngredientsRes = {
+    success: boolean;
+    data: Array<TIngredient>;
+  }
